[FIT-32] - Created endpoint routing for POST for sending Friend Requests
Created a routes file friend_request.js under routes/friends. This file creates endpoint routing for the POST request to send a friend request.

This adds the username of the user who sent the friend request, to the friend_request list of the receiver.

Edited the user schema to add friend_requests list attribute.