[S2-22] - Created POST endpoint to schedule workout
Edited schedule_workouts.js in routes/schedule.

Added a POST function called scheduleWorkouts and added the business logic that takes care of scheduling a workout for a user.

This function checks whether the user is authenticated, the workoutID passed in the req body is valid, and each of the friends username in the friend list is a valid friend. After all the checks, I created an object for schedule workout to be pushed to the list of scheduled workouts for the user and all the friends he has scheduled the workout with.

Added a route endpoint for this POST function in handler.js