[S3-29] - Created POST endpoint to set user's like status for workout
Updated the workout schema to add like and dislike attributes. These attributes are objects that consist of list of users (user id's) who have liked/disliked the workouts and the like/dislike count

Created the endpoint for the POST method to set user's like status for a workout

Created the POST method called updateLikeStatus that updates the like/dislike status of a workout by a user based on the workoutID and the status.

Tested the endpoint using a python script.